requirements.txt
speech_recognition
pip
RPi.GPIO 